Refactor integration tests for Redis backend to use base tests
